requirements.txt
requests
PyYAML
fastapi
uvicorn